requirement.txt
fastapi
uvicorn
openai
python-multipart
pillow
python-dotenv
tqdm
beautifulsoup4
requests


